requirements.txt
streamlit
fastapi
uvicorn
pytesseract
opencv-python
pillow
spacy
regex
pdfplumber
reportlab
